Refactor core media files into focused UI/navigation/services
