+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/media/navigation/MediaAppNavigation.kt
@@ -23,12 +23,7 @@ import com.universalmedialibrary.ui.media.player.*
 import com.universalmedialibrary.ui.media.screens.*
 import com.universalmedialibrary.ui.media.viewmodels.*
 import com.universalmedialibrary.debug.ui.DebugMenuScreen
-import com.universalmedialibrary.ui.components.NavigationItems
-import com.universalmedialibrary.ui.main.MainViewModel
-import com.universalmedialibrary.ui.theme.CleverFerretTheme
-import com.universalmedialibrary.ui.theme.ThemePalette
-import com.universalmedialibrary.ui.theme.toCleverFerretTheme
-import kotlinx.coroutines.launch
+import import kotlinx.coroutines.launch
 import com.universalmedialibrary.ui.reader.DocumentReaderScreen
 import com.universalmedialibrary.ui.reader.EPUBReaderScreen
 import java.io.File
@@ -280,283 +275,8 @@ fun MediaAppNavHost(
             )
         }
         
-        composable(MediaRoutes.SETTINGS) {
-            com.universalmedialibrary.ui.settings.SettingsScreen(
-                onBack = { navController.popBackStack() },
-                navController = navController,
-                availableBottomItems = NavigationItems.items
-            )
-        }
-        
-        // API Settings sub-screen
-        composable(MediaRoutes.SETTINGS_API) {
-            com.universalmedialibrary.ui.settings.APISettingsScreen(
-                onNavigateBack = { navController.popBackStack() }
-            )
-        }
+        addSettingsRoutes(navController)
 
-        // =========================
-        // Settings routes (fully operational - no placeholders)
-        // =========================
-
-        // Media-centric settings sub-pages (routes used by MediaSettingsScreen)
-        composable(MediaRoutes.SETTINGS_APPEARANCE) {
-            com.universalmedialibrary.ui.settings.AppearanceSettingsScreen(
-                onBack = { navController.popBackStack() },
-                navController = navController
-            )
-        }
-        composable("settings/social") {
-            com.universalmedialibrary.ui.settings.SocialIntegrationsSettingsScreen(
-                onBack = { navController.popBackStack() },
-                navController = navController
-            )
-        }
-        composable("settings/metadata") {
-            com.universalmedialibrary.ui.settings.MetadataSettingsScreen(
-                onBack = { navController.popBackStack() },
-                navController = navController
-            )
-        }
-        composable("settings/auto-scan") {
-            com.universalmedialibrary.ui.settings.AutoScanSettingsScreen(
-                onBack = { navController.popBackStack() }
-            )
-        }
-        composable("settings/display") {
-            com.universalmedialibrary.ui.settings.DisplaySettingsScreen(
-                onBack = { navController.popBackStack() },
-                navController = navController
-            )
-        }
-        composable("settings/text-size") {
-            // Reuse the existing reader settings page to adjust font sizes and layout.
-            com.universalmedialibrary.ui.settings.ReaderSettingsScreen(
-                navController = navController,
-                settingsType = "visual"
-            )
-        }
-        composable("settings/downloads") {
-            com.universalmedialibrary.ui.settings.DownloadsSettingsScreen(
-                onBack = { navController.popBackStack() }
-            )
-        }
-        composable("settings/cache") {
-            com.universalmedialibrary.ui.settings.CacheSettingsScreen(
-                onBack = { navController.popBackStack() }
-            )
-        }
-        composable("settings/privacy") {
-            com.universalmedialibrary.ui.settings.PrivacySettingsScreen(
-                onBack = { navController.popBackStack() }
-            )
-        }
-        composable("settings/casting") {
-            com.universalmedialibrary.ui.settings.CastingSettingsScreen(
-                onBack = { navController.popBackStack() }
-            )
-        }
-        composable("settings/feedback") {
-            com.universalmedialibrary.ui.settings.FeedbackSettingsScreen(
-                onBack = { navController.popBackStack() }
-            )
-        }
-
-        // Theme gallery (used by SettingsScreen)
-        composable("theme_showcase") {
-            val mainViewModel: MainViewModel = hiltViewModel()
-            val selectedTheme by mainViewModel.selectedTheme.collectAsState(ThemePalette.NAVY_GOLD)
-            val currentUnifiedTheme = selectedTheme.toCleverFerretTheme()
-
-            com.universalmedialibrary.ui.screens.ThemeShowcaseScreen(
-                currentTheme = currentUnifiedTheme,
-                onThemeSelected = { newTheme ->
-                    val oldPalette = when (newTheme) {
-                        CleverFerretTheme.NAVY_GOLD -> ThemePalette.NAVY_GOLD
-                        CleverFerretTheme.ROYAL_SILVER -> ThemePalette.ROYAL_SILVER
-                        CleverFerretTheme.FOREST_COPPER -> ThemePalette.FOREST_COPPER
-                        CleverFerretTheme.BURGUNDY_ROSE_GOLD -> ThemePalette.BURGUNDY_ROSE_GOLD
-                        CleverFerretTheme.CHARCOAL_CHAMPAGNE -> ThemePalette.CHARCOAL_CHAMPAGNE
-                        CleverFerretTheme.SLATE_GUNMETAL -> ThemePalette.SLATE_GUNMETAL
-                        else -> ThemePalette.NAVY_GOLD
-                    }
-                    mainViewModel.setTheme(oldPalette)
-                },
-                onNavigateBack = { navController.popBackStack() }
-            )
-        }
-
-        // Advanced audio effects showcase (used by AudioEffectsSettingsScreen)
-        composable("advanced_effects_showcase") {
-            com.universalmedialibrary.ui.screens.AdvancedEffectsShowcaseScreen()
-        }
-
-        // Import sorter (used by SettingsScreen)
-        composable("settings/import_sorter") {
-            com.universalmedialibrary.ui.settings.ImportSorterScreen(
-                onBack = { navController.popBackStack() }
-            )
-        }
-
-        // Import history (used by SettingsScreen)
-        composable("settings/import_history") {
-            com.universalmedialibrary.ui.settings.ImportHistoryScreen(
-                onBack = { navController.popBackStack() }
-            )
-        }
-
-        // Import/export (legacy settings route)
-        composable("settings/import_export") {
-            com.universalmedialibrary.ui.settings.ImportExportScreen(
-                onNavigateBack = { navController.popBackStack() }
-            )
-        }
-
-        // Media servers (used by SettingsScreen)
-        composable(MediaRoutes.SETTINGS_MEDIA_SERVERS) {
-            com.universalmedialibrary.ui.settings.MediaServerSettingsScreen(
-                onBack = { navController.popBackStack() }
-            )
-        }
-
-        // Back-compat alias (hyphenated route referenced by older media-centric settings UI)
-        composable("settings/media-servers") {
-            com.universalmedialibrary.ui.settings.MediaServerSettingsScreen(
-                onBack = { navController.popBackStack() }
-            )
-        }
-
-        // Network storage / cloud shares (used by SettingsScreen)
-        composable("settings/network_storage") {
-            com.universalmedialibrary.ui.settings.NetworkStorageSettingsScreen(
-                onBack = { navController.popBackStack() }
-            )
-        }
-
-        // Back-compat alias for media-centric route
-        composable("settings/cloud") {
-            com.universalmedialibrary.ui.settings.NetworkStorageSettingsScreen(
-                onBack = { navController.popBackStack() }
-            )
-        }
-
-        // Audio effects (used by SettingsScreen)
-        composable("settings/audio_effects") {
-            val vm: com.universalmedialibrary.ui.settings.AudioEffectsViewModel = hiltViewModel()
-            com.universalmedialibrary.ui.settings.AudioEffectsSettingsScreen(
-                viewModel = vm,
-                onNavigateBack = { navController.popBackStack() },
-                onNavigateToShowcase = { navController.navigate("advanced_effects_showcase") }
-            )
-        }
-
-        // Back-compat alias for media-centric route
-        composable(MediaRoutes.SETTINGS_PLAYBACK) {
-            val vm: com.universalmedialibrary.ui.settings.AudioEffectsViewModel = hiltViewModel()
-            com.universalmedialibrary.ui.settings.AudioEffectsSettingsScreen(
-                viewModel = vm,
-                onNavigateBack = { navController.popBackStack() },
-                onNavigateToShowcase = { navController.navigate("advanced_effects_showcase") }
-            )
-        }
-
-        // Audio profiles (used by SettingsScreen)
-        composable("settings/audio_profiles") {
-            com.universalmedialibrary.ui.settings.AudioProfilesScreen(
-                onNavigateBack = { navController.popBackStack() }
-            )
-        }
-
-        // Last.fm (used by SettingsScreen)
-        composable("settings/lastfm") {
-            com.universalmedialibrary.ui.settings.LastFmSettingsScreen(
-                onNavigateBack = { navController.popBackStack() }
-            )
-        }
-
-        // Text-to-speech providers (used by SettingsScreen)
-        composable("settings/tts_provider") {
-            com.universalmedialibrary.ui.settings.TtsProviderSettingsScreen(
-                onBack = { navController.popBackStack() }
-            )
-        }
-
-        // Back-compat alias for media-centric route
-        composable("settings/tts") {
-            com.universalmedialibrary.ui.settings.TtsProviderSettingsScreen(
-                onBack = { navController.popBackStack() }
-            )
-        }
-
-        // Parental controls (used by SettingsScreen)
-        composable("settings/parental_controls") {
-            com.universalmedialibrary.ui.settings.ParentalControlsScreen(
-                navController = navController
-            )
-        }
-
-        // Back-compat alias for media-centric route
-        composable("settings/parental") {
-            com.universalmedialibrary.ui.settings.ParentalControlsScreen(
-                navController = navController
-            )
-        }
-
-        // Storage organizer (legacy settings route)
-        composable("settings/organizer") {
-            com.universalmedialibrary.ui.settings.StorageOrganizerScreen(
-                onBack = { navController.popBackStack() }
-            )
-        }
-
-        // Back-compat alias for media-centric route
-        composable(MediaRoutes.SETTINGS_STORAGE) {
-            com.universalmedialibrary.ui.settings.StorageOrganizerScreen(
-                onBack = { navController.popBackStack() }
-            )
-        }
-
-        // Reader settings (wire to existing ReaderSettingsScreen section entry)
-        composable(MediaRoutes.SETTINGS_READER) {
-            com.universalmedialibrary.ui.settings.ReaderSettingsScreen(
-                navController = navController,
-                settingsType = "visual"
-            )
-        }
-
-        // Security (wire to operational security settings)
-        composable(MediaRoutes.SETTINGS_SECURITY) {
-            com.universalmedialibrary.ui.settings.SecuritySettingsScreen(
-                navController = navController
-            )
-        }
-
-        // About
-        composable(MediaRoutes.SETTINGS_ABOUT) {
-            com.universalmedialibrary.ui.settings.AboutScreen(
-                navController = navController
-            )
-        }
-
-        // Ambient utilities (used by SettingsScreen)
-        composable("ambient/theme_manager") {
-            com.universalmedialibrary.ui.ambient.ThemeManagerScreen(
-                onBack = { navController.popBackStack() }
-            )
-        }
-        composable("ambient/import") {
-            com.universalmedialibrary.ui.ambient.AudioPackImportScreen(
-                onBack = { navController.popBackStack() }
-            )
-        }
-
-        // Fanfiction hub (used by SettingsScreen)
-        composable("fanfiction_hub") {
-            com.universalmedialibrary.ui.webfiction.UnifiedFanfictionHubScreen(
-                navController = navController
-            )
-        }
-        
         // =====================================================================
         // LIBRARY SCREENS
         // =====================================================================
