+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/settings/SettingsPrimarySections.kt
@@ -1,0 +1,198 @@
+package com.universalmedialibrary.ui.settings
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Switch
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.navigation.NavController
+import com.universalmedialibrary.ui.theme.MetallicButton
+import com.universalmedialibrary.ui.theme.MetallicCard
+import com.universalmedialibrary.ui.theme.MetallicDivider
+import com.universalmedialibrary.ui.theme.MetallicText
+
+@Composable
+internal fun AppearanceSettingsSection(
+    uiState: SettingsUiState,
+    navController: NavController,
+    onOpenThemePicker: () -> Unit,
+    onDarkModeChanged: (Boolean) -> Unit,
+    onReduceMotionChanged: (Boolean) -> Unit
+) {
+    MetallicText(
+        text = "Appearance",
+        style = MaterialTheme.typography.titleMedium,
+        modifier = Modifier.padding(vertical = 8.dp)
+    )
+
+    MetallicCard {
+        Column {
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(vertical = 8.dp),
+                horizontalArrangement = Arrangement.SpaceBetween,
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Column(modifier = Modifier.weight(1f)) {
+                    Text(
+                        text = "Theme",
+                        style = MaterialTheme.typography.titleSmall,
+                        fontWeight = FontWeight.Medium
+                    )
+                    Text(
+                        text = uiState.selectedTheme.name.replace("_", " "),
+                        style = MaterialTheme.typography.bodySmall,
+                        color = MaterialTheme.colorScheme.onSurfaceVariant
+                    )
+                }
+
+                Row(horizontalArrangement = Arrangement.spacedBy(8.dp)) {
+                    MetallicButton(
+                        text = "Gallery",
+                        onClick = { navController.navigate("theme_showcase") }
+                    )
+                    MetallicButton(
+                        text = "Change",
+                        onClick = onOpenThemePicker
+                    )
+                }
+            }
+
+            MetallicDivider()
+
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(vertical = 8.dp),
+                horizontalArrangement = Arrangement.SpaceBetween,
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Column(modifier = Modifier.weight(1f)) {
+                    Text(
+                        text = "Dark Mode",
+                        style = MaterialTheme.typography.titleSmall,
+                        fontWeight = FontWeight.Medium
+                    )
+                    Text(
+                        text = "Use dark theme for better reading",
+                        style = MaterialTheme.typography.bodySmall,
+                        color = MaterialTheme.colorScheme.onSurfaceVariant
+                    )
+                }
+
+                Switch(
+                    checked = uiState.darkMode,
+                    onCheckedChange = onDarkModeChanged
+                )
+            }
+
+            MetallicDivider()
+
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(vertical = 8.dp),
+                horizontalArrangement = Arrangement.SpaceBetween,
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Column(modifier = Modifier.weight(1f)) {
+                    Text(
+                        text = "Reduce Motion",
+                        style = MaterialTheme.typography.titleSmall,
+                        fontWeight = FontWeight.Medium
+                    )
+                    Text(
+                        text = "Turn off auto-scrolling and animated transitions",
+                        style = MaterialTheme.typography.bodySmall,
+                        color = MaterialTheme.colorScheme.onSurfaceVariant
+                    )
+                }
+
+                Switch(
+                    checked = uiState.reduceMotion,
+                    onCheckedChange = onReduceMotionChanged
+                )
+            }
+        }
+    }
+}
+
+@Composable
+internal fun PodcastSettingsSection(
+    uiState: SettingsUiState,
+    onAutoDownloadChanged: (Boolean) -> Unit,
+    onWifiOnlyChanged: (Boolean) -> Unit
+) {
+    MetallicText(
+        text = "Podcasts",
+        style = MaterialTheme.typography.titleMedium,
+        modifier = Modifier.padding(top = 16.dp, bottom = 8.dp)
+    )
+
+    MetallicCard {
+        Column {
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(vertical = 8.dp),
+                horizontalArrangement = Arrangement.SpaceBetween,
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Column(modifier = Modifier.weight(1f)) {
+                    Text(
+                        text = "Auto-Download Episodes",
+                        style = MaterialTheme.typography.titleSmall,
+                        fontWeight = FontWeight.Medium
+                    )
+                    Text(
+                        text = "Automatically download new episodes",
+                        style = MaterialTheme.typography.bodySmall,
+                        color = MaterialTheme.colorScheme.onSurfaceVariant
+                    )
+                }
+
+                Switch(
+                    checked = uiState.autoDownloadPodcasts,
+                    onCheckedChange = onAutoDownloadChanged
+                )
+            }
+
+            MetallicDivider()
+
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(vertical = 8.dp),
+                horizontalArrangement = Arrangement.SpaceBetween,
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Column(modifier = Modifier.weight(1f)) {
+                    Text(
+                        text = "WiFi Only Downloads",
+                        style = MaterialTheme.typography.titleSmall,
+                        fontWeight = FontWeight.Medium
+                    )
+                    Text(
+                        text = "Download only on WiFi to save data",
+                        style = MaterialTheme.typography.bodySmall,
+                        color = MaterialTheme.colorScheme.onSurfaceVariant
+                    )
+                }
+
+                Switch(
+                    checked = uiState.wifiOnlyDownloads,
+                    onCheckedChange = onWifiOnlyChanged
+                )
+            }
+        }
+    }
+}
